feat: Settings backend controller
